Feature Development(Mobile Movement): Enable touch and swipe controlls for mobile users
Added new booleans to weaponTeleport script, which make sure tapping and swiping are recognised.
Added new if statements to weaponTeleport script. They take a swipe input and move the weapon in the swiped direction. Upon cancel or end of swipe the weapon is thrown.
Added a Standalone Region and a Mobile Region to clean up the code and differ between PC and Mobile.
Disabled custom cursor and normal cursor.