Fix(Anim Player & WallJump.sc): Player transitions
Added transitions between the different animations: idle, jump, fall.
Added bools to check for the conditions of each. Changed the box
colliders of the borders on the camera, they are now bigger. Changed the
WallJump.sc so that it now checks for falling, jumoing and idle and also
whether to flip the sprite render on the X axis.

Sprites_Update #3 W.W.